fix(contracts): upgrade scholarship_fund to soroban-sdk 25.x
Change ScholarshipApplication.status from u8 to u32; Soroban #[contracttype] only supports 32-bit and wider integers.

+++ contracts/scholarship_fund/src/lib.rs
@@ -18,7 +18,7 @@ pub struct ScholarshipApplication {
     pub id: u64,
     pub student: Address,
     pub amount_requested: i128,
-    pub status: u8, // 0=pending, 1=approved, 2=rejected, 3=distributed
+    pub status: u32, // 0=pending, 1=approved, 2=rejected, 3=distributed
     pub applied_at: u64,
 }
